requirements.txt
numpy>=1.26
scipy>=1.11
scikit-learn>=1.4
tqdm>=4.66
pyyaml>=6.0
torch>=2.2
wandb>=0.17
